Loading paths with accents : isolate "normalize" in a separate lib + normalize file result writer
