+++ src/libs/antares/study/area/list.cpp
@@ -29,9 +29,9 @@
 #include <antares/study/area/scratchpad.h>
 #include "antares/antares/antares.h"
 #include "antares/array/matrix.h"
+#include "antares/paths/normalize_paths.h"
 #include "antares/study//study.h"
 #include "antares/study/area/area.h"
-#include "antares/study/normalize_paths.h"
 #include "antares/study/parts/load/prepro.h"
 #include "antares/study/parts/parts.h"
 #include "antares/utils/utils.h"
