+++ src/libs/antares/study/parts/common/cluster.cpp
@@ -24,7 +24,7 @@
 
 #include <antares/study/study.h>
 #include <antares/utils/utils.h>
-#include "antares/study/normalize_paths.h"
+#include "antares/paths/normalize_paths.h"
 
 namespace fs = std::filesystem;
