+++ src/libs/antares/study/parts/hydro/prepro.cpp
@@ -27,7 +27,7 @@
 
 #include <antares/inifile/inifile.h>
 #include <antares/logs/logs.h>
-#include "antares/study/normalize_paths.h"
+#include "antares/paths/normalize_paths.h"
 #include "antares/study/study.h"
 
 using namespace Antares;
