+++ src/libs/antares/study/xcast/xcast.cpp
@@ -26,7 +26,7 @@
 #include <antares/inifile/inifile.h>
 #include <antares/logs/logs.h>
 #include <antares/utils/utils.h>
-#include "antares/study/normalize_paths.h"
+#include "antares/paths/normalize_paths.h"
 #include "antares/study/study.h"
 
 using namespace Yuni;
